Add level dimension moving mechanic
Make the parallel dimension above the player represent the end of the level
at the start of the level, finishing with the start of the level when the player
reaches the end of the level in the main dimension.

To make the top dimension move with the player, make the top parallel move
with the player's relative z distance from the start point.